refactor(dir): Atualizado estrutura de pastas
