+++ src/main.ts
@@ -1,24 +1,25 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app/app.module';
+import { AppModule } from './app.module';
 import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
-import { LoggingInterceptor } from './logging/logging.interceptor';
-import './config'
+import { LoggingInterceptor } from './common/interceptors/logging.interceptor';
+import './config';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
 
   const config = new DocumentBuilder()
-  .setTitle('API de Clientes')
-  .setDescription('A API de gerenciamento de clientes.')
-  .setVersion('1.0') 
-  .addBearerAuth()
-  .build();
+    .setTitle('API de Clientes')
+    .setDescription('A API de gerenciamento de clientes.')
+    .setVersion('1.0')
+    .addBearerAuth()
+    .build();
 
   const document = SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('api', app, document);
 
   app.useGlobalInterceptors(new LoggingInterceptor());
 
+  app.enableCors();
   await app.listen(process.env.PORT ?? 3000);
 }
 bootstrap();
