+++ src/modules/auth/interfaces/auth.service.interface.ts
