+++ src/modules/client/entities/client.entity.ts
